+++ EssentialFeed/EssentialFeedTests/Feed Cache/Helpers/FeedCacheTestHelpers.swift
@@ -1,0 +1,30 @@
+//
+//  FeedCacheTestHelpers.swift
+//  EssentialFeedTests
+//
+//  Created by Aaron Huánuco on 13/09/2020.
+//  Copyright © 2020 Aaron Huánuco. All rights reserved.
+//
+
+import Foundation
+import EssentialFeed
+
+func uniqueImage() -> FeedImage {
+    FeedImage(id: UUID(), description: "any", location: "any", url: anyURL())
+}
+
+func uniqueImageFeed() -> (models: [FeedImage], local: [LocalFeedImage]) {
+    let models = [uniqueImage(), uniqueImage()]
+    let local = models.map { LocalFeedImage(id: $0.id, description: $0.description, location: $0.location, url: $0.imageURL)}
+    return (models, local)
+}
+
+extension Date {
+    func adding(days: Int) -> Date {
+        Calendar(identifier: .gregorian).date(byAdding: .day, value: days, to: self)!
+    }
+
+    func adding(seconds: TimeInterval) -> Date {
+        self + seconds
+    }
+}
