+++ EssentialFeed/EssentialFeedTests/FeedAPI/URLSessionHTTPClientTests.swift
@@ -131,9 +131,7 @@ final class URLSessionHTTPClientTests: XCTestCase {
         return receivedResult
     }
 
-    private func anyURL() -> URL { URL(string: "https://any-url.com")!}
     private func anyData() -> Data { Data("any data".utf8) }
-    private func anyNSError() ->  NSError {  NSError(domain: "any error", code: 0) }
     private func nonHTTPURLResponse() -> URLResponse {
         URLResponse(url: anyURL(), mimeType: nil, expectedContentLength: 0, textEncodingName: nil)
     }
